Substantial refactoring in preparation for training: - Remove unneccessary files - Remove data folder to reflect the nonlocality of our data - Move game_converter and preprocessing out of models, into new folder called preprocessing - Rename training scripts for consistency
