--- AlphaGo/models/shallow_policy.py
@@ -1,1 +1,0 @@
-# TODO
