+++ AlphaGo/preprocessing/sgflib/README.txt
