+++ AlphaGo/preprocessing/sgflib/lgpl.txt
