+++ AlphaGo/preprocessing/sgflib/typelib.py
