+++ AlphaGo/training/supervised_policy_trainer.py
